Replace in-tree sha256 implementation with openssl wrapper
Not using our own cryptography will be important to any vendors who
need to certify compliance (e.g., with FIPS-140-2).  Since we already
depend on openssl, just replace the embedded tomcrypt with that.

To keep the nicer libtomcrypt interface, add a thin wrapper in
openssl.c around the openssl EVP interface.  The only API changes
are (1) that the sha_ctx type changes and (2) that sha256_done() can
fail (which is unfortunate but unavoidable).

Support versions of openssl both older and newer than 1.1.0.  This is
a bit gross since the underlying type of a digest context changes
between them, and also because the way they are allocated changes its
name.  These details are hidden in the provided wrapper.